requirements.txt
fastapi
uvicorn
mysql-connector-python
jinja2
gspread
oauth2client
